design fixes and changes
