+++ lib/Cards/LessonCard.dart
@@ -3,6 +3,7 @@ import 'package:e_szivacs/generated/i18n.dart';
 import 'package:flutter/material.dart';
 
 import '../Datas/Lesson.dart';
+import '../globals.dart' as globals;
 import 'dart:async';
 
 class LessonCard extends StatelessWidget {
@@ -167,50 +168,59 @@ class LessonCard extends StatelessWidget {
       },
       child: new Card(
         margin: EdgeInsets.all(6.0),
-        child: new Column(
-          children: <Widget>[
-            new Container(
-              child: Wrap(
-                children: <Widget>[
-                  new Text(
-                    S.of(context).next_lesson,
-                    style: new TextStyle(
-                      fontSize: 18.0,
-                    ),
-                  ),
-                  new Text(getNext() != null ? getNext().subject : "error",
-                      style: new TextStyle(
-                          fontSize: 18.0, color: Colors.blueAccent)),
-                  new Text(", ",
+        child: new Container(
+          child: new Column(
+            children: <Widget>[
+              new Container(
+                child: Wrap(
+                  children: <Widget>[
+                    new Text(
+                      S.of(context).next_lesson,
                       style: new TextStyle(
                         fontSize: 18.0,
-                      )),
-                  Container(
-                    padding: EdgeInsets.only(right: 5),
-                    child: new Text(
-                        getRemainingTime(
-                            S.of(context).hour, S.of(context).minute),
+                      ),
+                    ),
+                    new Text(getNext() != null ? getNext().subject : "error",
                         style: new TextStyle(
                             fontSize: 18.0, color: Colors.blueAccent)),
-                  ),
-                  new Text(
-                    S.of(context).later,
-                    style: new TextStyle(
-                      fontSize: 18.0,
+                    new Text(", ",
+                        style: new TextStyle(
+                          fontSize: 18.0,
+                        )),
+                    Container(
+                      padding: EdgeInsets.only(right: 5),
+                      child: new Text(
+                          getRemainingTime(
+                              S.of(context).hour, S.of(context).minute),
+                          style: new TextStyle(
+                              fontSize: 18.0, color: Colors.blueAccent)),
                     ),
-                    softWrap: false,
-                    maxLines: 2,
-                  ),
-                ],
-                alignment: WrapAlignment.start,
+                    new Text(
+                      S.of(context).later,
+                      style: new TextStyle(
+                        fontSize: 18.0,
+                      ),
+                      softWrap: false,
+                      maxLines: 2,
+                    ),
+                  ],
+                  alignment: WrapAlignment.start,
+                ),
+                alignment: Alignment(-1, 0),
+                padding: EdgeInsets.all(10.0),
+                color: globals.isDark ? Color.fromARGB(255, 25, 25, 25) : Colors.grey[300],
               ),
-              alignment: Alignment(-1, 0),
-              padding: EdgeInsets.all(10.0),
-            ),
-            new Divider(
-              height: 1.0,
-            ),
-            new Container(
+              new Container(
+                decoration: ShapeDecoration(
+                  shape: RoundedRectangleBorder(
+                    side: BorderSide(
+                      style: BorderStyle.none,
+                      width: 1,
+                    ),
+                    borderRadius: BorderRadius.circular(4),
+                  ),
+                  color: globals.isDark ? Color.fromARGB(255, 15, 15, 15) : Colors.white,
+                ),
                 padding: EdgeInsets.all(10.0),
                 child: new Padding(
                   padding: new EdgeInsets.all(0.0),
@@ -229,7 +239,15 @@ class LessonCard extends StatelessWidget {
                     ],
                   ),
                 ))
-          ],
+            ],
+          ),
+        decoration: new BoxDecoration(
+          border: Border.all(
+            color: globals.isDark ? Color.fromARGB(255, 25, 25, 25) : Colors.grey[300],
+            width: 2.5),
+          borderRadius: new BorderRadius.all(Radius.circular(5)),
+          color: globals.isDark ? Color.fromARGB(255, 25, 25, 25) : Colors.grey[300],
+          ),
         ),
       ),
     );
